requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aioredis-cluster==1.5.2
aiosignal==1.2.0
alibabacloud-credentials==0.3.2
alibabacloud-dyvmsapi20170525==2.1.4
alibabacloud-ecs20140526==3.0.7
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.2.1
alibabacloud-rocketmq20220801==1.0.5
alibabacloud-tea==0.2.9
alibabacloud-tea-openapi==0.3.7
alibabacloud-tea-util==0.3.8
alibabacloud-tea-xml==0.0.2
aliyun-python-sdk-alidns==2.6.32
aliyun-python-sdk-cdn==3.7.1
aliyun-python-sdk-core==2.13.26
aliyun-python-sdk-cs==4.8.1
aliyun-python-sdk-domain==3.14.7
aliyun-python-sdk-dyvmsapi==3.2.0
aliyun-python-sdk-ecs==4.19.11
aliyun-python-sdk-elasticsearch==3.0.29
aliyun-python-sdk-kms==2.15.0
aliyun-python-sdk-nas==3.14.0
aliyun-python-sdk-ons==3.1.8
aliyun-python-sdk-polardb==1.8.13
aliyun-python-sdk-r-kvstore==2.20.2
aliyun-python-sdk-rds==2.5.1
aliyun-python-sdk-slb==3.3.9
aliyun-python-sdk-waf-openapi==1.1.8
amqp==2.6.1
ansible==2.9.13
arrow==1.2.3
asgiref==3.4.1
asttokens==2.0.8
async-timeout==4.0.2
atlassian-python-api==3.36.0
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
autopep8==2.0.2
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.4.0
blessed==1.20.0
boto3==1.18.32
botocore==1.21.65
cachetools==4.1.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
channels==3.0.4
channels-redis==3.2.0
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
croniter==1.3.8
cryptography==3.4.8
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2
django-auth-ldap==2.4.0
django-celery-beat==2.0.0
django-cors-headers==3.5.0
django-excel==0.0.10
django-fernet-fields==0.6
django-filter==2.3.0
django-mirage-field==1.4.0
django-oauth-toolkit==2.1.0
django-picklefield==3.1
django-q==1.3.9
django-redis==4.12.1
django-rest-framework-condition==0.1.1
django-timezone-field==4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
dnspython==2.4.2
docutils==0.15.2
drf-nested-routers==0.93.4
drf-social-oauth2==1.2.1
drf-yasg==1.20.0
ecdsa==0.18.0
elasticsearch==7.15.1
elasticsearch-dsl==7.4.0
et-xmlfile==1.0.1
executing==1.0.0
frozenlist==1.3.1
google-auth==1.22.1
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==4.3.0
incremental==17.5.0
inflection==0.5.1
ipython==8.5.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.1
Jinja2==2.11.2
jira==3.4.1
jmespath==0.10.0
jwcrypto==1.3.1
kombu==4.6.11
kubernetes==17.17.0
ldap3==2.9.1
lml==0.1.0
Markdown==3.3.5
MarkupSafe==1.1.1
matplotlib-inline==0.1.6
msgpack==1.0.2
multi-key-dict==2.0.3
multidict==6.0.2
mybatis-mapper2sql==0.1.9
mysqlclient==2.0.1
numpy==1.26.1
oauthlib==3.1.0
openpyxl==3.1.2
oss2==2.15.0
packaging==20.9
pandas==2.1.1
paramiko==2.7.2
parso==0.8.3
pbr==5.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.31
psutil==5.8.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.20
pycryptodome==3.9.8
pyexcel==0.6.7
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
Pygments==2.13.0
PyHamcrest==2.0.2
PyJWT==2.5.0
pymongo==3.11.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python-gitlab==3.15.0
python-jenkins==1.7.0
python-jose==3.3.0
python-ldap==3.3.1
python-redis-lock==3.7.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.31.0
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
rest-social-auth==7.0.0
rsa==4.5
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
s3transfer==0.5.2
SchemaObject==0.5.9
semantic-version==2.10.0
service-identity==18.1.0
setuptools-rust==1.5.2
shortuuid==1.0.1
simplejson==3.18.3
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
stack-data==0.5.0
tencentcloud-sdk-python==3.0.529
texttable==1.6.4
tomli==2.0.1
tqdm==4.65.0
traitlets==5.3.0
Twisted==21.2.0
txaio==20.4.1
typing==3.7.4.3
typing-extensions==3.10.0.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
websocket-client==0.57.0
wrapt==1.14.1
XlsxWriter==3.1.2
xlwt==1.3.0
xmltodict==0.12.0
yarl==1.8.1
zipp==3.4.1
zope.interface==5.1.2
orderedset==2.0.3
supervisor==4.2.5
